Function combining functions (contractions) given as parameters into a list of class IFS.

Function combining functions (contractions) given as parameters into a list of class IFS.

createIFS(f = NULL, ...)

Arguments

f
First function on a list.

...
Remaining functions on a list.

Value

A list of class IFS.

Examples

lin3 <- function(a,b) {function(x,y){c(1/3*x+a/3,1/3*y+b/3)}}
createIFS(lin3(0,0),lin3(0,1),lin3(0,2),lin3(1,0),lin3(1,2),lin3(2,0),lin3(2,1),lin3(2,2))

#> [[1]]
#> function (x, y) 
#> {
#>     c(1/3 * x + a/3, 1/3 * y + b/3)
#> }
#> <environment: 0x10c40c4a8>
#> 
#> [[2]]
#> function (x, y) 
#> {
#>     c(1/3 * x + a/3, 1/3 * y + b/3)
#> }
#> <environment: 0x10c40c240>
#> 
#> [[3]]
#> function (x, y) 
#> {
#>     c(1/3 * x + a/3, 1/3 * y + b/3)
#> }
#> <environment: 0x10c240f80>
#> 
#> [[4]]
#> function (x, y) 
#> {
#>     c(1/3 * x + a/3, 1/3 * y + b/3)
#> }
#> <environment: 0x10c240d18>
#> 
#> [[5]]
#> function (x, y) 
#> {
#>     c(1/3 * x + a/3, 1/3 * y + b/3)
#> }
#> <environment: 0x10c240ab0>
#> 
#> [[6]]
#> function (x, y) 
#> {
#>     c(1/3 * x + a/3, 1/3 * y + b/3)
#> }
#> <environment: 0x10c2417f0>
#> 
#> [[7]]
#> function (x, y) 
#> {
#>     c(1/3 * x + a/3, 1/3 * y + b/3)
#> }
#> <environment: 0x10c241588>
#> 
#> [[8]]
#> function (x, y) 
#> {
#>     c(1/3 * x + a/3, 1/3 * y + b/3)
#> }
#> <environment: 0x10c241320>
#> 
#> attr(,"class")
#> [1] "list" "IFS"